--- sedaro/src/sedaro/models/workspace/WorkspaceManager.py
@@ -1,10 +1,0 @@
-from typing import ClassVar
-
-from ..BaseModelManager import BaseModelManager
-from .Workspace import Workspace
-
-
-class WorkspaceManager(BaseModelManager[Workspace]):
-
-    _BASE_PATH: 'ClassVar[str]' = '/workspaces'
-    _MODEL = Workspace
